mobilenet-infer/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
requests
